WIP: Implement the Directory Layer
This is surprisingly complicated for all that this layer does. The
DirectoryPartition feature/concept in particular throws a fairly hefty
wrench into the Erlang implementation due to a lack of OO concepts that
are quite heavily relied on.

As such while this implementation does attmept to match the Python
version in terms of naming and general behavior, the implementation is
a fair bit more "clunky" due to the lack of syntax extensions.